DomainsConcertadas

+++ SpMedicalGroup.WebAPI/SpMedicalGroup.WebAPI/Domains/Consultum.cs
@@ -9,9 +9,9 @@ namespace SpMedicalGroup.WebAPI.Domains
     {
         public int IdConsulta { get; set; }
         public int IdPaciente { get; set; }
-        public int IdMedico { get; set; }
+        public short IdMedico { get; set; }
         public string DataConsulta { get; set; }
-        public int IdSituacao { get; set; }
+        public byte IdSituacao { get; set; }
 
         public virtual Medico IdMedicoNavigation { get; set; }
         public virtual Paciente IdPacienteNavigation { get; set; }
