+++ SpMedicalGroup.WebAPI/SpMedicalGroup.WebAPI/Domains/Medico.cs
@@ -13,10 +13,10 @@ namespace SpMedicalGroup.WebAPI.Domains
             Usuarios = new HashSet<Usuario>();
         }
 
-        public  int IdMedico { get; set; }
+        public short IdMedico { get; set; }
         public string Crm { get; set; }
-        public int? IdClinica { get; set; }
-        public int? IdEspecialidade { get; set; }
+        public byte? IdClinica { get; set; }
+        public byte? IdEspecialidade { get; set; }
 
         public virtual Clinica IdClinicaNavigation { get; set; }
         public virtual Especialidade IdEspecialidadeNavigation { get; set; }
